refactor(api): move HTTP request protocols into shared nodes protocols module
